Add minimial example of fsspec>=2022 failure
tox allows easier testing accross versions

the problem is caused by handling of the gs:// prefix by _strip_protocol
