+++ tests/conftest.py
@@ -5,6 +5,7 @@ from fsspec.implementations.memory import MemoryFileSystem
 import fv3config.filesystem
 import fv3config.data
 from . import mocks
+from .mocks import MockGCSFileSystem
 
 TEST_DIR = os.path.dirname(os.path.realpath(__file__))
 
@@ -21,7 +22,10 @@ DIAG_TABLE_PATH = "gs://vcm-fv3config/config/diag_table/default/v1.0/diag_table"
 
 def populate_mock_filesystem(fs):
     for filename in MOCK_FS_FILENAMES:
-        fs.mkdir(os.path.dirname(filename))
+
+        parent_dir = os.path.dirname(filename)
+        if not fs.exists(parent_dir):
+            fs.mkdir(parent_dir)
         with fs.open(filename, mode="wb") as f:
             f.write(b"mock_data")
     with open(
@@ -29,31 +33,6 @@ def populate_mock_filesystem(fs):
     ) as f_in:
         with fs.open(DIAG_TABLE_PATH, "w") as f_out:
             f_out.write(f_in.read())
-
-
-class MockGCSFileSystem(MemoryFileSystem):
-
-    protocol = "gcs", "gs"
-
-    def ls(self, path, recursive=False, **kwargs):
-        path = self._strip_protocol(path)
-        return super().ls(path, **kwargs)
-
-    def mkdir(self, path, create_parents=True, **kwargs):
-        path = self._strip_protocol(path)
-        return super().mkdir(path, create_parents, **kwargs)
-
-    def rmdir(self, path, *args, **kwargs):
-        path = self._strip_protocol(path)
-        return super().rmdir(path, *args, **kwargs)
-
-    def exists(self, path):
-        path = self._strip_protocol(path)
-        return path in self.store or path in self.pseudo_dirs
-
-    def open(self, path, *args, **kwargs):
-        path = self._strip_protocol(path)
-        return super().open(path, *args, **kwargs)
 
 
 @pytest.fixture(autouse=True)
