+++ tox.ini
@@ -1,0 +1,19 @@
+[tox]
+envlist = {un,}constrained
+
+[testenv:{constrained,unconstrained}]
+install_command = 
+    constrained: pip install -c constraints.txt {opts} {packages}
+    unconstrained: pip install {opts} {packages}
+allowlist_externals = rm
+    mkdir
+deps = pytest
+    constrained: -cconstraints.txt
+setenv =
+    FV3CONFIG_CACHE_DIR={toxinidir}/.cache
+commands = 
+    
+    rm -rf {toxinidir}/.cache
+    mkdir -p {toxinidir}/.cache
+    pytest -k 'not kubernetes'
+
